One upstream commit since cc1f5c2:

- #469 (cab110e) GLM-4's router correction bias is declared and loaded as
  fp32 instead of inheriting the engine's bf16. Upcasting it at use cannot
  recover the rounding, which can change which experts a token near a top-k
  boundary selects. Upstream validated it on synthetic weights on CPU; the
  disagreement rate on a real checkpoint is unmeasured.

No conflicts; Kai does not touch models/glm4_moe. Kai's
_materialize_loaded_weight_state_dict casts each loaded tensor to its model
buffer's dtype, so the fp32 declaration is what lands. The bias is a resident
weight, so the --moe-bank-ram and offload banks never see it, and no model
served on either machine takes this path.

+++ python/freetoken/models/glm4_moe/moe.py
@@ -32,10 +32,8 @@ class Glm4MoeSparseBlock(BaseOP):
         self.topk_group = config.topk_group
 
         self.gate = LinearReplicated(config.hidden_size, config.num_experts, has_bias=False)
-        # DeepSeek-style selection bias; a registered buffer in HF (kept fp32 there). We
-        # store it in the model's bf16 dtype and upcast at use, which is exact enough for
-        # the argmax-style top-k selection.
-        self.e_score_correction_bias = torch.empty(config.num_experts)
+        # Keep selection bias in fp32: rounding can change the selected experts.
+        self.e_score_correction_bias = torch.empty(config.num_experts, dtype=torch.float32)
 
         # The offload cache indexes experts by *MoE* layer (global layer minus
         # first_k_dense_replace), matching how the loader packs the expert banks. The
